Updating Views With One to Many Join Can Cause Surprising Results
Be especially careful when you allow modifications against a view defined by a join query. Users who are not aware that the target object for their modifications is a view and not a table might find the effect of their modifications surprising in some cases -- for example, when they modify the "one" side of a one-to-many join.